Add method to attempt to find original attachment
We might have attachments in the database with the incorrect hexdigest.
If so we won't be able to find the original attachment body. This commit
adds a method which accounts for 2 scenarios which might cause the
hexdigest to change. 1. different line endings, 2 leading zeros on dates

Both of these changes have been caused by upstream changes to the mail
gem we use to parse the emails but this hasn't been apparent until
recently.