Add admin toolbox section to public body list
It's still too easy to use the mass tag manipulation form as a search
box. We do now have a confirmation alert, but it's still annoying.

This commit adds a "toolbox" section so that infrequently used tools can
be hidden away until needed.

I did look at using the Bootstrap collapse [1] element, but it wasn't
quite right as I only wanted one heading that collapses the whole
section. The Bootstrap component added lots of extra markup, so went
with a simple `<summary>` tag for an initial implementation. We can
tweak it further if other tools get added.

[1] https://getbootstrap.com/2.2.0/javascript.html#collapse