Fix embedded attachment processing
For attachments which are embedded within a RFC822 attachment, were
unable process some attachment if they were received a long time ago.

This is down to changes in the underlying mail gem which results in
slight difference in the headers which get appended to the attachment
body.

This change fixes this by only matching against the body of embedded
attachments, ignoring headers and any slight changes.

Fixes #7876