Fix attachment processing encoding issues
Switch to binary unsafe line end conversion helpers. This fixes an issue
where attachments couldn't be masked correctly due to encoding, line
endings and the presence of non-ASCII characters.

When we have different encodings, say UTF-8 attachment from the raw
email and an 8BIT binary data from Active Storage, the safe line end
conversion helpers won't operate when the binary data has non-ASCII
characters [1].

This meant in the hex digest never matching and the original unmasked
attachment body not being retrievable. With this change this is fixed.

Using the unsafe conversion helpers is fine in this case as the strings
are only used to generate MD5 hex digests.

[1] https://github.com/mikel/mail/blob/10a4443/lib/mail/utilities.rb#L248-L250

Fixes #7880