Add ability to ask external caches to invalidate.
Add a cached_urls method to the Comment, FoiAttachment, InfoRequest,
PublicBody and User models, containing the entries that those models
wish to invalidate either when an event is logged, via log_event, or
when the censor rules change, or when a body or user is updated. The
entries are fixed paths to be purged, or regexes beginning with ^ to
be banned.

Adds a NotifyCacheJob that for each of the object's cached_urls, for
each locale, for each host in VARNISH_HOSTS, makes BAN or PURGE HTTP
requests to the host, to invalidate the relevant entries.