Fix uuencoded attachment processing
For attachments which are encoded with Unix-to-Unix encoding we are
unable process some attachments in order to apply masks or censor rules
to them.

This is down to when the attachments are extracted and saved into the
database. It happens outside of our `MailHandler` library which we're
using to retrieve the original unmasked attachment body for processing.

This change fixes this by checking plain text attachments for begin/end
markers and calculating the relevant hexdigest for comparison.

Fixes #7877