Add method to catch RebuiltAttachment exceptions
When calling `FoiAttachment#body` this can now raise an exception due to
the current attachment instance getting replaced and no longer existing
in the database.

We can catch and rescue from this by reloading the new attachment and
calling the same code again.

This is what `FoiAttachment.protect_against_rebuilt_attachments` does
and should prevent exceptions when viewing requests which haven't be
masked and cached yet. These will normally be very old attachments as
this also need the `hexdigest` to be mismatched.