Update attachment masking before action callbacks
We're going to make a change which might mean attachments might be
rebuilt within the `FoiAttachmentMaskJob` this means the signed global
ID might not return the attachment in this controller. This change
ensures their is a referer parameter and if the attachment can't be
loaded then we redirect back to the referer.

Note, we don't need to worry about people modifying the referer to get
users redirected to a third party domain as this is protected against by
Rails.