Remove attachment prominence check
This check isn't doing anything. It was meant to protect when parsing
incoming messages to stop hidden attachments from being made public
again.

We don't actually yet have this parsing in place instead we've focused
on finding the original attachment [1].

For the remainder of cases we will need to parse the raw email again but
we have added separate check [2] so this one here is redundant.

[1] https://github.com/mysociety/alaveteli/issues/7875
[2] https://github.com/mysociety/alaveteli/pull/7920