Refactor RequestMailer#receive
Loop through exact matches and delivery emails to each info request.
This is now possible because we check `InfoRequest#already_received?`
when calling `InfoRequest#receive`.

Previously this could result in requests receiving an email multiple
times due to the same email coming in to the mail server multiple times,
once for each address in the `To` header.

Fixes #7913