Add a basic TrackThing both_links
We render Tracks in the admin interface but they don't have an "eye"
next to them as per other records. This commit adds a basic
implementation which just links to the search query of the track.

The current view code in `admin_track/_some_tracks` does some more
complex handling, where the "title" of the track links directly to a
trackable record if the track has one associated, or falls back to the
plain text search query.

We could later move this behaviour to the eye, and then link to an admin
page for the track via the search term (in line with other records) when
we have an admin page.

The main intent of this commit is just to get the eye in place, such
that we can improve the behaviour at a later date.