Show tag-based notes on new request preview
Use TaggableTerms to check OutgoingMessage#body for terms that we have a
tag-based Note for.

We can use this to warn users that they’re about to make a mistake.

For example, we could configure OutgoingMessage like so:

    OutgoingMessage.taggable_terms = {
      body: {
        /my 999 call/i => ‘preview_warning_sar’
      }
    }

Then, if a user includes “my 999 call” in their request content, we’d
detect this and render a Note applied via the ‘preview_warning_sar’ tag.