Update AttachmentMaskController referer handling
Addresses potential cross-site scripting issue identified by Brakeman.

This change uses `MessageVerifier` to encodes the referer when passing
it from the `AttachmentController` to the `AttachmentMaskController`,
this is needed because we don't know which route the attachment was
requested from (public, share by link, project).

We were relying on `redirect_to` raising an `UnsafeRedirectError`
exception when the referer param was changes to an external link.

This works for the controller, but we also use the referer param on the
`done` view. So once the attachment is available, if the param is
manually changed the resulting page would have a download link which
points to changed referer param.

We use temporary signed attachment IDs so the resulting URL would only
be available for a short period of time so the attack vector is minimal.