Update "show as HTML" attachment controller action
This updates `AttachmentController#show_as_html` to do attachment
masking in the background.

Our "view attachments in the browser" feature, or "view as HTML", is
doing a lot of processing inline in the web processes and this sometimes
causes memory issues on the server and results in oom-kill running.

Uses the same logic as the `#show` action to check an attachment is
masked before proceeding with the action. This means it will time outs
after 5 seconds and then redirect the `AttachmentMaskController` to wait
for the job to complete before rendering a view saying the attachment is
now viewable.

Fixes #7977