Add Github workflow to check for changelog entry
When creating PR we should update the changelog as this helps others see
what is upcoming in the next release and makes release Alaveteli
versions easier.

This workflow ensures an `docs/CHANGES.md` is modified or the pull
request description contains `[skip changelog]`

Update the PR template to help remind devs to do this before the PR is
created.