Update attachment processing
Handle when incoming messages are re-parsed which can cause attachments
to be rebuilt when trying to render the main body part.

Occasionally when main body attachment part will get rebuilt while the
content is masked. When this happens we can see errors such as:
`undefined method `unmasked_body' for nil:NilClass`.

This change ensures the masking job doesn't fail and the code will now
raise an `MissingAttachment` exception instead.

Rescuing from this in `IncomingMessage#get_main_body_text_internal`
allows us to retry straight away as in the second attempt the main part
will be the new attachment instance and the part will be rendered.

Fixes #7971