Admin view of unmasked incoming message
Admins often want to view the contents of an incoming message without
any masking or censor rules applied.

In some cases you can inspect the raw email, but it's often base64
encoded, which means needing to download the email and open it in a mail
client. Even when the email is not base64 encoded, it's often difficult
to mentally parse the information out when there are lots of headers,
attachment sections and HTML body content.

This commit adds an unmasked view of main body part attachments. This is
equivalent to the main body contents of an incoming message that we show
in public.

This does require "downloading" the attachment from Active Storage, but
at present we use local file storage for attachments, so that's fine.
Even if we moved them to S3 it wouldn't be a disaster to fetch them each
time, as this isn't a frequently viewed page.

Needed to use an `!important` CSS rule to improve the visual rendering
as for some reason bootstrap sets `margin-bottom: 0;` on `p` tags within
a `blockquote` element (See
bootstrap-sass-2.3.2.2/vendor/assets/stylesheets/bootstrap/_type.scss).

Fixes https://github.com/mysociety/alaveteli/issues/1106.