Add new Category model
This model uses nested sets model technique to store hierarchy and the
ordering of categories.

Adds the same associations from existing `PublicBodyCategory` model
which this model will be replacing to retain backwards compatibly for
now.

As `PublicBodyHeading` will be migrated into to `Category` instances we
have dropped the validation on the `description` and `category_tag`
attributes.