Improve clarity of both_links for FoiAttachment
Prepending the info request title adds extra noise when viewing the list
of an attachments for an incoming message. We lose a bit of context if
rendering `both_links` for an attachment outside of its request/incoming
message page, but I don't think we do that at present and it's worth the
tradeoff. In a situation where we do want both, then we can add it in in
that specific context.