Fix install script if Ruby version changes
If a install upgrades Ruby re-running the install scripts will fail due
to the newer Ruby version not being present.

This happens when re-running as there would be a `config/general.yml`
file present and we're checking the `STAGING_SITE` configuration is
consistent with the install script flags.

Setting `RBENV_VERSION` to `system` uses the Debian/Ubuntu packages Ruby
to run `bin/config` to get `STAGING_SITE`.