Fix exception loading attachment unmasked body
When loading raw unmasked attachments from storage this can sometimes
fail due to the stored files being missing. On WDTK this seems to
happens sometimes with old files which weren't in the cache when we
migrated to `ActiveStorage`.

We always used to guard against missing files in the old files cache but
since the attachment processing refactor this hasn't happened.

This change brings this back but using the same `MissingAttachment`
exception used to rebuild attachments from raw emails when the original
record in the database can't be found (EG that iswhen the hexdigest has
changed).