Fix exception loading attachment masked body
When loading masked attachments from storage we still occasionally see
failures due to the stored files being missing.

We always used to guard against missing files in the old files cache but
since the attachment processing refactor this hasn't happened.

This change simply rescues the exception allowing the attachment to be
re-masked and the missing file to be recreated from the raw email.