Update FoiAttachment#body
Allow attachments which have been rebuilt to find the body from the new
attachment record.

Previously this would only happen if `#persisted?` is true but this
method checks to see if an instance has been marked for destruction.

This meant the loading of the new attachment could only ever happen
once.

Moving outside this conditional fixes exceptions seen in development
where the hexdigest is invalid and the stored file on disk has been
removed. Its not impossible this could happen in production also.