Integrate ActionMailbox
Integrate ActionMailbox into `RequestMailer.receive` for inbound email
handling. This maintains existing functionality but lays the groundwork
for future enhancements.

Added an `origin` column to the ActionMailbox::InboundEmail table. This
change was necessary to preserve the functionality of checking the mail
`source`, avoiding a naming clash with the `source` method.

Benefits of using ActionMailbox include:

1. Enhanced email routing capabilities, enabling specialized
   processing like the Excel hidden data spreadsheet analyzer.
2. Refactoring opportunities for `RequestMailer#receive`, particularly
   for spam detection, duplicate email handling, and initial request
   assessment.
3. Clear separation of concerns between Mailers (for sending) and
   Mailboxes (for receiving).
4. Improved email processing efficiency through ActionMailbox,
   facilitating background job handling and potential simplification
   of mail ingress.
5. Provides a solution for re-users to receive emails without needing
   their own mail server setup.