Improve redirection flow for user closure actions
In practice we often use all three closure actions in a single step –
Close; Anonymise; Erase. It's annoying to get taken back to the show
page and then have to go back in to edit to take the next action, and so
on.

This commit returns you to the edit page for the Close and Anonymise
actions. Once an account gets erased, you're returned to the show page
to signify the "end" of the possible actions.

Fixes https://github.com/mysociety/alaveteli/issues/7986.