Refactor: Generalise login method on language
This uses CSS selectors rather than labels, so it will work with the
webpage being in other languages.

Also needed to prefix route in an integration test to get the correct
locale rendered, this is because the `signin_path` now doesn't specify
the locale so the session object hasn't been set.