Fix storing session locale
When the locale changes we need to ensure the users session is updated.
Normally this would happen using the session object but we have a
middleware to strip "empty" sessions updates, this includes the locale
key. This appears to be left over from an optimisation for Varnish
cache. See [1] and [2].

Storing a different key makes the session update work. This is important
for this feature as we will be removing the prefixed locale in the URL
and we have to ensure this is set correctly.

[1] https://github.com/mysociety/alaveteli/commit/fe895b0
[2] https://chris.heald.me/2011/rails-varnish-cookie-sessions-and-csrf-tokens/