Add Note#rich_body
Use ActionText Trix editor to allow rich text notes. We render the rich
text body when the note isn't in the original style.

This commit also hides the image attachment button in the Trix editor
as we will need to store the files correctly which won't work currently
with the current ActiveStorage setup.