Refactor request list routes
Remove routes for "all" and "recent" views.

The "recent" view already is redirected in the main `list` action to the
"all" view.

The "all" view should really be the default for the `list` action as
currently the `/list` endpoint will show no requests because there is no
view is selected.