Tweak Note.sort behaviour
The default mental model for notes is roughly analogous to the Bootstrap
alert levels [1]:

* Red: Error/Danger
* Yellow: Warning
* Green: Success
* Blue: Information

This commit tweaks the ordering such that notes are sorted as above. If
there's some sort of "Danger" note, then we want to show it before more
general "Information".

This ordering won't _always_ be correct, but it'll be Good Enough for
the majority of cases.

This also fixes the associated test. The `match_array` expectation is
order-agnostic [2]:

    expect([1, 2, 3]).to match_array([2, 3, 1]) # pass

[1] https://getbootstrap.com/2.2.0/components.html#alerts
[2] https://rspec.info/features/3-12/rspec-expectations/built-in-matchers/contain-exactly/