Update admin tag routing constraints
Allow admin pages for tags with `.` or `/` to be loaded successfully.
This means tags can be URLs which is currently a use case used on WDTK.

It does mean we can't have member actions on the tags resource but as of
now we don't have any. If ever we want an `edit` action we can adjust
the constraint to allow this.