Update Xapian indexing health check
Modify the health check to track the oldest job in the queue more
effectively.

Previously, the health check inspected the created_at timestamp of the
oldest job, which could be misleading if many jobs were queued
simultaneously.

Now, during each check, we will determine whether the oldest job in the
queue has changed. We will log the current time whenever a change is
detected and continue to return this time until we observe a change in
the oldest job.

This ensures the health check will only fail if the job queue is stuck,
with no jobs processed for over an hour.