Fix assignment of event associations
If we're logging an event where an associated object has been deleted
the event won't be able to created due to record being missing.

This change rescues the exception when the record isn't found and
assigns the record params to what would've been the global ID. This
allows us visibility that there was an object associated at one point in
time.