Add some links to /citations
Add some basic navigation links to the new public list of citations.

It would be nice to add a homepage section for this by extracting out
the intro text of citations/index into a partial, but let's leave that
for a later release to give people a chance to build up a list of
citations before heavier promotion.