Add test storage service
In Rails 7.1 it will require a default ActiveStorage service to be
defined.

This adds a service for the test environment so test examples
don't break when we bump the Rails gem.

This commit ensures the tests can be run on future bump commit. The
other environments will be defined when we run `rails app:update`.