Update locale loading before app initialisation
Sometimes we need locales to be loaded before the application has been
initialised.

Setting in the in `config.before_configuration` block doesn't persist
correctly and also needs to be called in after initialisation.

This issue becomes apparent when there is an exception at boot.

The error message might call `to_sentence` [1], which will attempt to
load a translation and this then error if `set_locales` hasn't been
called.

[1]: https://github.com/rails/rails/blob/d39db5d/activestorage/lib/active_storage/service/registry.rb#L18-L19

See #5382