Configure regexp timeout
Rails 8.0 set the default to 1 second. We're not ready for that yet.

We have been hitting this in production and although the c12dc6d helped
with the majority of cases, particularly when masking attachments in the
background queue.

There are still instances where we see this exception. E.g., Xapian
indexing and RequestController#show action when viewing messages which
haven't been masked as we try to do this inline in the web request.

Once both of these have been moved to the background, we can go back to
the original fix.