Migrate away from annotate gem
This gem is unmaintained and has been replaced by annotaterb. This works
slightly differently so this commit setups the new gem while replicating
as closely as possible the setup we previously had.

There will be some minor changes to the annotations in the next couple
of commits to take into account differences in how the new gem outputs.