Add report link on user profile page
The UK Online Safety Act requires good content reporting capabilities.

We have report links attached to most user generated content, but not
here. This is a quick fix to add a report link to user profile pages,
which goes to the general contact form.

In future it would be good to have this direct to a more structured
form, but this works for now as we consider user profiles more
generally.