Move dataset download button
Moves the project dataset download button above the table. Projects can
include hundreds of requests, which results in a very long table. This
makes it easy to miss that you can download the dataset as a CSV, or if
you do know, makes for a very long scroll to get there.