Update InfoRequest cache expiry process
Don't expire attachment mask if the attachment is locked. This would
potentially break as the masking process will need to go back to the raw
email to retrieve the unmasked attachment body.

This might not exist if the raw email has been removed or this might be
incorrect if the attachment has been replaced via the admin UI.