Update IncomingMessage attachment extraction process
This prevents the accidental deletion of locked attachments. We can't
rebuild locked attachments from the raw email as it could be deleted or
the attachment replaced via the admin UI.

We don't handle the exceptions here anywhere yet and exception mails
will be sent. If this is a common problem we can look at how to prevent
the attachment extraction if an incoming message is locked.