Add attachment replacement to admin edit form
Allow admins to replace an attachment. This will also automatically lock
the attachment.

The purpose of uploading a replacement is to allow admins to redact
attachments offsite using other software.

The main body part, which should be text content can be edited inline in
the admin form to make replacement easier.

Also record of an attachment has been replaced in the edit_attachment
event.