Fix attachment locking removing redacted filenames
When locking an attachment which has a censor rule applied to the
attachment's filename we need to retain the redacted filename otherwise,
after its been locked, the rule won't be applied and the redaction
fails.

This meant the attachment locking feature isn't always useable and
admins would always need to replace attachments instead.