Hide the search form at the top of a category results page
As noted by @gbp in #8445, the search form at the top of the category
(search results) pages takes up a lot of space, and isn’t very useful,
since if you’ve arrived at the page, you already know what you’re
interested in. We don‘t want to remove the search form entirely,
though, as it might be useful to power users looking to narrow the
tagged results.

This commit moves the category description _above_ the search form, and
wraps the form in a `<details>` tag, so that it’s hidden by default but
still usable if people need it.

Fixes the final part of #8445.