Ability to limit response upload permissions
In some cases we may wish to exclude any user registered with an email
address within the authorities domain from responding as an FOI officer.

This is most useful for organisations with a very large staff, where
domains are shared between staff and others (e.g. education
institutions), or where the authority uses a more generic domain for
their FOI address.

In future it would be good to be able to add regexp to the list (to e.g.
exclude all `.ac.uk` domains).