Extract PublicBody::CsvImport concern
Simply move the methods to a concern given they're an isolated concept
and take up a significant amount of line length in the main class.

Obviously there's lots of refactoring to be done here – ideally this
would be a well-tested class in its own right with the concern being a
simple interface, but for now just extracting the pieces out to their
own file is a nice improvement.