Add rate limiting to InfoRequest
Prevent users using their daily limit at unreasonable speeds.

We don't want to allow a user to sign up and smash through their daily
limit in 5 minutes, then create a new account to do the same, and so on.

This ensures that the allocation is used sensibly. It hopefully won't
affect the majority of genuine users.