Remove User#next_request_permitted_at
We're going to introduce rate limiting within the daily cap, so the next
request time will be much more difficult to calculate. This would add a
fair amount of complexity for a minor value-add. In addition, most
genuine users shouldn't hit the rate limit within the daily cap, so we
don't really want to give misusers additional information to allow them
to work around the rate limiter.