Update cattr_reader methods with defaults
When the default value itself is defined as a class attribute the
default value can be assigned to these class attributes before any
overrides are defined in the Alaveteli theme.

This change overwrites the class reader method to return the default
value but it is not assigned to the class attribute and will remain set
to nil if not assigned to anything in the Alaveteli theme.

To save having to also override the instance reader method in the
same way as the class reader method, this change also updates any
instance calls to the attributes.