Fix signup button layout
Other form fields are within a <p> tag to give them margin bottom. In
WDTK we customise this view to add another form_item_note underneath the
button, which gets jammed up against it.

A <p> isn't the most ideal way to fix this but it mirrors what's there
without having to mess around with the whole view.