Extract admin_incoming_message/holding_pen_actions
Create a partial for actions relevant to the holding pen. The other
actions in `admin_incoming_message/actions` are irrelevant here, so this
makes the holding pen view much clearer and makes the holding pen
related conditional a bit more intentional.

Also removes the guess handling from the normal actions since that's
only relevant in the holding pen.