Refactor Users::ConfirmationsController
This change unifies actions for consistency.

All actions for the different post redirect circumstances are basically
doing the same thing.

The only difference is for the 'change_password' action. If we take into
account the change password controller/action which we get redirected to
these actually perform the same operations just in different orders. EG
before this change, the user would be sign in after password change form
is submitted, now they get signed in before being redirected to the
form.

Additionally the `@user` var was being used in two different contexts;
the current logged in user and the post redirect's user, so this
refactor clears that up also.