Update project queue submission variable
This will allow the same views to be used for the show/edit actions.

As the submission instance variable will be pass to views we can direct
for form to create/update actions without specifying the post/patch
method type, and render different content based on if the submission is
persisted or not.