Update project submission redirects
This change ensures the user is redirect back to where they came from.
For example if coming from the project dataset view redirect back there.

This would work correctly when the edit/update actions are hit but if a
request has never been classified or had data extracted, the submission
resource hasn't been persisted and we'll hit the create action.