Update users:destroy_unused rake task
Use the `last_sign_in_at` timestamp to ensure we don't destroy users
who've signed in to keep their account active.

On WDTK we sent notifications to users with unused accounts prompting
them to sign in to keep their account. This change ensure we won't
delete their account when we enable the cronjob.